Create hygeia medic box
The HYGEIA MEDIC BOX is a smart medic box , It is used especially for the old age people who forget the place where they kept their medical box . The hygeia medic box is designed in such a way that if a person missed their medical box , They should make a call to the sim ejected in the medic box. The buzzer sound will alarmed from the missing medic box and corresponding led light are glowed from the medic box. The led lights are specially designed for deaf people. With the help of the buzzer sound and led the missed medical box can be found