Create smart dustbin
The Smart Dustbin Project introduces an automated waste disposal system using Arduino technology. The system utilizes ultrasonic sensors and a servo motor to detect the fill level of a dustbin and automatically open its lid when it reaches a certain capacity.

Through the integration of an Arduino microcontroller, ultrasonic sensor, and servo motor, the project offers a simple yet effective solution for optimizing waste management processes. By continuously measuring the distance between the lid and the trash inside the bin, the system ensures timely lid opening to facilitate waste disposal.

The Arduino code employs pulse echo principles to measure distance accurately, while an averaging mechanism enhances precision. When the measured distance indicates that the bin is nearly full, the servo motor is activated to open the lid for a specific duration, allowing users to deposit their waste conveniently.

This project presents an accessible and cost-efficient approach to improving waste collection efficiency, which can be implemented in various settings to promote cleanliness and sustainability.